+++ mara_acl/views.py
@@ -87,7 +87,12 @@ def save_permissions():
 
 @mara_acl.before_app_request
 def login():
+    # exclude some uris from login
+    if (any(flask.request.path.startswith(uri) for uri in config.whitelisted_uris())):
+        return None
+
     email = flask.request.headers.get(config.email_http_header())
-    response = users.login(email if email else 'guest@localhost')
-    if response != True:
-        return response
+    if not users.login(email if email else 'guest@localhost'):
+        flask.abort(403, 'Hi ' + email + ',<br/><br/>We haven\'t created an account for you yet. '
+                    + 'Please contact the person that gave you the link this site to fix that.')
+