+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 *2017-03-10* 
 
 - Initial version
+
+## 1.0.1
+*2017-04-15*
+
+- Refactored login 
+- Added function `user_has_permission`